// src/types/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export interface Database {
  graphql_public: {
    Tables: Record<never, never>;
    Views: Record<never, never>;
    Functions: {
      graphql: {
        Args: {
          extensions?: Json;
          operationName?: string;
          query?: string;
          variables?: Json;
        };
        Returns: Json;
      };
    };
    Enums: Record<never, never>;
    CompositeTypes: Record<never, never>;
  };
  public: {
    Tables: {
      trips: {
        Row: {
          dropoff_coordinates: unknown | null;
          dropoff_datetime: string | null;
          id: number;
          passenger_count: number | null;
          pickup_coordinates: unknown | null;
          pickup_datetime: string | null;
          store_and_fwd_flag: string | null;
          suspicious_trip: boolean | null;
          trip_duration: number | null;
          trip_min_distance: number | null;
          trip_speed: number | null;
          vendor_id: number | null;
        };
        Insert: {
          dropoff_coordinates?: unknown | null;
          dropoff_datetime?: string | null;
          id?: number;
          passenger_count?: number | null;
          pickup_coordinates?: unknown | null;
          pickup_datetime?: string | null;
          store_and_fwd_flag?: string | null;
          suspicious_trip?: boolean | null;
          trip_duration?: number | null;
          trip_min_distance?: number | null;
          trip_speed?: number | null;
          vendor_id?: number | null;
        };
        Update: {
          dropoff_coordinates?: unknown | null;
          dropoff_datetime?: string | null;
          id?: number;
          passenger_count?: number | null;
          pickup_coordinates?: unknown | null;
          pickup_datetime?: string | null;
          store_and_fwd_flag?: string | null;
          suspicious_trip?: boolean | null;
          trip_duration?: number | null;
          trip_min_distance?: number | null;
          trip_speed?: number | null;
          vendor_id?: number | null;
        };
        Relationships: [
          {
            foreignKeyName: "trips_vendor_id_fkey";
            columns: ["vendor_id"];
            isOneToOne: false;
            referencedRelation: "vendors";
            referencedColumns: ["id"];
          },
        ];
      };
      vendors: {
        Row: {
          id: number;
          name: string | null;
        };
        Insert: {
          id?: number;
          name?: string | null;
        };
        Update: {
          id?: number;
          name?: string | null;
        };
        Relationships: [];
      };
    };
    Views: Record<never, never>;
    Functions: Record<never, never>;
    Enums: Record<never, never>;
    CompositeTypes: Record<never, never>;
  };
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  "public"
>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never;

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never;